Move file format types to separate file.
